pyproject.toml
[tool.poetry]
name = "django_zarinpal"
version = "0.1.6"
description = "integrate django payments with zarrinpal"
authors = ["glyphack <sh.hooshyari@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.5"
zeep = "^3.4.0"
django-hashid-field = "^3.1.1"

[tool.poetry.dev-dependencies]
pycodestyle = "^2.6.0"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
